Revert "Feature: added reducers and notifications"
