+++ backend/micro/actions/createAccount.js
@@ -18,6 +18,6 @@ export default async ({ email, password }, mongo) => {
       created: new Date().getTime(),
     });
 
-    return {};
+    return true;
   }
 };
