--- frontend/nextjs/components/Login/index.js
@@ -1,48 +1,0 @@
-import { useState } from 'react';
-import useStore from '../../hooks/useStore';
-import { login, createAccount } from './functions';
-
-export default () => {
-  const dispatch = useStore()[1];
-  const [attemptLogin, setAttemptLogin] = useState(true);
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
-  const onChangeEmail = (e) => setEmail(e.target.value);
-  const onChangePassword = (e) => setPassword(e.target.value);
-
-  const toggleAttemptLogin = (e) => {
-    e.preventDefault();
-
-    setAttemptLogin(!attemptLogin);
-  };
-
-  const action = attemptLogin
-    ? login(email, password, dispatch)
-    : createAccount(email, password, dispatch);
-
-  return (
-    <div>
-      <form onSubmit={action}>
-        <input
-          type='text'
-          name='email'
-          onChange={onChangeEmail}
-          value={email}
-        />
-        <input
-          type='password'
-          name='password'
-          onChange={onChangePassword}
-          value={password}
-        />
-        <button onClick={action}>
-          {attemptLogin ? 'Login' : 'Create account'}
-        </button>
-      </form>
-      <a onClick={toggleAttemptLogin}>
-        Go to {attemptLogin ? 'Create account' : 'Login'}
-      </a>
-    </div>
-  );
-};
