--- frontend/nextjs/components/Logout/functions.js
@@ -1,9 +1,0 @@
-import cookie from 'js-cookie';
-import Router from 'next/router';
-
-const logout = () => {
-  cookie.remove('token');
-  Router.push('/');
-};
-
-export { logout };
