+++ frontend/nextjs/package.json
@@ -12,14 +12,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "js-cookie": "^2.2.1",
     "lodash": "^4.17.15",
     "next": "^9.3.5",
-    "next-cookies": "^2.0.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "sha1": "^1.1.1",
-    "short-uuid": "^3.1.1",
     "socket.io-client": "^2.3.0"
   }
 }
