+++ frontend/nextjs/pages/index.js
@@ -1,11 +1,3 @@
 import Login from '../components/Login';
-import Logout from '../components/Logout';
-import auth from '../common/auth';
 
-const Index = ({ token }) => (token ? <Logout token={token} /> : <Login />);
-
-Index.getInitialProps = (ctx) => ({
-  ...auth(ctx),
-});
-
-export default Index;
+export default () => <Login />;
